Split ui into separate subpackage
Move LinuxDesktopParser into ui subpackage, it does not belong in the
Opensnitch core
