+++ opensnitch/connection.py
@@ -23,7 +23,7 @@ from socket import inet_ntoa, getservbyport
 
 
 class Connection:
-    def __init__(self, packet_id, procmon, desktop_parser, payload):
+    def __init__(self, packet_id, procmon, payload):
         self.id = packet_id
         self.data     = payload
         self.pkt      = ip.IP( self.data )
@@ -66,8 +66,7 @@ class Connection:
                                                             self.dst_addr,
                                                             self.dst_port,
                                                             self.proto)
-            self.app = Application(procmon, desktop_parser,
-                                   self.pid, self.app_path)
+            self.app = Application(procmon, self.pid, self.app_path)
             self.app_path = self.app.path
 
     def get_app_name(self):
