+++ opensnitch/ui/resources/dialog.ui
