+++ opensnitch/ui/resources/dialog_hi.ui
